+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 - Revert the steppers to the font setting controls in the Appearance settings pane.
 - Remove the Script menu from the context menu.
+- Update the Swift syntax to support highlighting regex literals.
 
 
 ### Fixes