+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Revert the steppers to the font setting controls in the Appearance settings pane.
 - Remove the Script menu from the context menu.
 - Add a help tag for “Show Sidebar” menu item when the command is unavailable.
+- Update the CSS syntax to add new keywords (thanks to Karam!).
 - Update the Swift syntax to support highlighting regex literals.
 - Update the JSON syntax to include the .xcstrings file extension.